Added dynamic room size functionality
Rooms can now be created that are larger than the 28x16 size. If width is a multiple of 30 (minus 2) or height is a multiple of 18 (minus 2), the room will be placed like any other except it contains multiple sectors on the map. Image must still be rectangular, however (so no L-shaped rooms).